getopts setup

+++ src/FileCacheCleaner.php
@@ -1,8 +1,8 @@
 <?php
 /**
  * File Cache Cleaner
- * Delete expired Laravel-style `Illuminate\Cache` cache files
- *
+ *  - Delete expired Laravel-style `Illuminate\Cache` cache files
+ *  - https://github.com/attogram/file-cache-cleaner
  */
 declare(strict_types = 1);
 
@@ -16,8 +16,10 @@ use RecursiveIteratorIterator;
 
 use function array_reverse;
 use function file_get_contents;
+use function getopt;
 use function get_class;
 use function gmdate;
+use function in_array;
 use function is_dir;
 use function preg_match;
 use function print_r;
@@ -30,10 +32,17 @@ use function unlink;
 class FileCacheCleaner
 {
     /** @var string Code Version */
-    const VERSION = '2.2.0';
+    const VERSION = '2.3.0';
 
     /** @var string Date Format for gmdate() */
     const DATE_FORMAT = 'Y-m-d H:i:s';
+
+    /** @var string Usage */
+    const USAGE = "Attogram File Cache Cleaner Usage:\n"
+    . " file-cache-cleaner --directory cacheDirectory --clean\n"
+    . "  Options:\n"
+    . "  -d path  or  --directory path  - set path to Cache Directory\n"
+    . "  -c       or  --clean           - clean cache: delete expired files, remove empty subdirectories\n\n";
 
     /** @var string $cacheDirectory - top-level of Cache Directory to be cleaned */
     private $cacheDirectory = '';
@@ -44,51 +53,63 @@ class FileCacheCleaner
     /** @var int $currentTime - current datetime in unix timestamp format */
     private $currentTime = 0;
 
-    /** @var mixed $verbose - verbosity level, empty = off, not-empty = on*/
-    private $verbose = '';
+    /** @var bool $clean - clean cache directory? */
+    private $clean = false;
 
-    /** @var array $counts - status counts */
-    private $count = [];
+    /** @var array $report - report on cache status */
+    private $report = [];
 
     /**
-     * @param string $directory (default '')
-     * @param mixed $verbosity - verbosity level (default '' off) empty = off, not-empty = on
+     * Clean The Cache Directory - delete expired cache files and empty directories
      */
-    public function clean(string $directory = '', $verbosity = '')
+    public function clean()
     {
-        $this->verbose = $verbosity;
-        $this->currentTime = time();
         $this->debug(get_class() . ' v' . self::VERSION);
-        $this->debug('Check Time: ' . gmdate(self::DATE_FORMAT, $this->currentTime));
-    
-        $this->setCacheDirectory($directory);
-        $this->debug('Cache Directory: ' . $this->cacheDirectory);
 
-        $this->count['files'] = $this->count['directories']
-            = $this->count['deleted_files'] = $this->count['deleted_dirs'] = 0;
-
+        $this->setOptions();
         $this->examineCacheDirectory();
-        $this->debug($this->count['files'] . ' cache files found');
-        $this->debug($this->count['directories'] . ' sub-directories found');
-        $this->debug($this->count['deleted_files'] . ' deleted cache files');
+        $this->examineCacheSubdirectories();
+
+        $this->debug($this->report);
+        
+    }
+
+    private function setOptions()
+    {
+        $options = getopt('d:c', ['directory:', 'clean']);
+
+        if (!$options) {
+            print self::USAGE;
+            $this->fatalError('Please specify --directory and --clean');
+        }
+
+        // -d or --directory - set Cache Directory
+        $cacheDirectory = !empty($options['d']) ? $options['d'] : '';
+        $cacheDirectory = !empty($options['directory']) ? $options['directory'] : $cacheDirectory;
+        $this->setCacheDirectory($cacheDirectory);
+
+        // -c or --clean - turn on Cleaning mode
+        $this->clean = isset($options['c']) ? true : false;
+        $this->clean = isset($options['clean']) ? true : $this->clean;
     
-        $this->examineDirectories();
-        $this->debug($this->count['deleted_dirs'] . ' deleted empty directories');
+        $this->currentTime = time();
     }
 
     /**
      * @param string $directory (default '')
-     * @throws InvalidArgumentException
      */
     private function setCacheDirectory(string $directory = '')
     {
         if (!$directory) {
-            throw new InvalidArgumentException('Missing Cache Directory');
+            print self::USAGE;
+            $this->fatalError('Missing Cache Directory. Please specify with -d or --directory');
         }
         if (!is_dir($directory)) {
-            throw new InvalidArgumentException('Cache Directory Not Found');
+            print self::USAGE;
+            $this->fatalError('Cache Directory Not Found');
         }
         $this->cacheDirectory = realpath($directory);
+        $this->debug('Cache Directory: ' . $this->cacheDirectory);
     }
 
     private function examineCacheDirectory()
@@ -99,6 +120,7 @@ class FileCacheCleaner
             RecursiveIteratorIterator::SELF_FIRST
         );
         foreach ($filesystemIterator as $splFileInfo) {
+            $this->incrementReport('objects');
             $this->examineObject($splFileInfo);
         }
     }
@@ -110,30 +132,44 @@ class FileCacheCleaner
     {
         // Find Illuminate\Cache files - filenames are 40 character hexadecimal sha1 hashes
         if ($splFileInfo->isFile() && strlen($splFileInfo->getFileName()) == 40) {
-            $this->count['files']++;
             $this->examineFile($splFileInfo->getPathName());
             
             return;
         }
-        // Save directories to list
+        // Save subdirectories to list
         if ($splFileInfo->isDir()) {
-            $this->count['directories']++;
+            $this->incrementReport('subdirectories');
             $this->subDirectoryList[] = $splFileInfo->getPathName();
+
+            return;
         }
+
+        $this->incrementReport('non_cache_files');
     }
 
     /**
-     * @param string $pathname - full path and filename
+     * @param string $pathname - path and filename
      */
     private function examineFile(string $pathname)
     {
-        if (!($timestamp = $this->getFileCacheExpiration($pathname)) // If no valid timestamp found
-            || ($timestamp >= $this->currentTime) // If file cache is Not Expired yet
-        ) {
+        if (!$timestamp = $this->getFileCacheExpiration($pathname))  { // If no valid timestamp found
+            $this->incrementReport('invalid_cache_expiration_files');
             return;
         }
+
+        if ($timestamp >= $this->currentTime) { // If file cache is Not Expired yet
+            $this->incrementReport('unexpired_cache_files');
+            return;
+        }
+
+        $this->incrementReport('expired_cache_files');
+
+        if (!$this->clean) {
+            return;
+        }
+
         if (unlink($pathname)) {
-            $this->count['deleted_files']++;
+            $this->incrementReport('deleted_cache_files');
     
             return;
         }
@@ -141,7 +177,7 @@ class FileCacheCleaner
     }
 
     /**
-     * @param string $pathname - full path and filename
+     * @param string $pathname - path and filename
      * @return int - expiration time as unix timestamp, or 9999999999 on error
      */
     private function getFileCacheExpiration(string $pathname): int
@@ -153,7 +189,8 @@ class FileCacheCleaner
             || strlen($timestamp) != 10 // if timestamp is Not 10 characters long
             || !preg_match('/^([0-9]+)$/', $timestamp) // if timestamp is Not numbers-only
         ) {
-            $this->debug('Not cache: ' . $pathname);
+            $this->incrementReport('non-cache-files');
+
             return 9999999999; // max time 2286-11-20 17:46:39
         }
 
@@ -163,11 +200,14 @@ class FileCacheCleaner
     /**
      * Remove Empty Directories
      */
-    private function examineDirectories()
+    private function examineCacheSubdirectories()
     {
         foreach (array_reverse($this->subDirectoryList) as $directory) {
             if ($this->isEmptyDirectory($directory)) {
-                $this->removeDirectory($directory);
+                $this->incrementReport('empty_directories');
+                if ($this->clean) {
+                    $this->removeDirectory($directory);
+                }
             }
         }
     }
@@ -195,7 +235,7 @@ class FileCacheCleaner
     private function removeDirectory($directory)
     {
         if (rmdir($directory)) {
-            $this->count['deleted_dirs']++;
+            $this->incrementReport('deleted_dirs');
             
             return;
         }
@@ -203,12 +243,56 @@ class FileCacheCleaner
     }
 
     /**
-     * @param mixed $msg
+     * @param string $key
+     * @param mixed $value
+     */
+    private function setReport($key, $value)
+    {
+        $this->report[$key] = $value;
+    }
+
+    /**
+     * Increment report value
+     * @param string $key
+     */
+    private function incrementReport($key)
+    {
+        if (empty($this->report[$key])) {
+            $this->report[$key] = 1;
+
+            return;
+        }
+
+        $this->report[$key]++;
+    }
+
+    /**
+     * @param string $key
+     * @return mixed
+     */
+    private function getReport($key)
+    {
+        if (isset($this->report[$key])) {
+            return $this->report[$key];
+        }
+
+        return 0;
+    }
+
+    /**
+     * @param mixed $msg (optional)
      */
     private function debug($msg = '')
     {
-        if ($this->verbose) {
-            print gmdate(self::DATE_FORMAT) . ' UTC: ' . print_r($msg, true) . "\n";
-        }
+        print gmdate(self::DATE_FORMAT) . ' UTC: ' . print_r($msg, true) . "\n";
+    }
+
+    /**
+     * @param mixed $msg (optional)
+     */
+    private function fatalError($msg = '')
+    {
+        print gmdate(self::DATE_FORMAT) . ' UTC: FATAL ERROR: ' . print_r($msg, true) . "\n";
+        exit;
     }
 }
